refactor(bitcoin-core-sv2): add v32x_v31x_v30x and v32x_v31x shared modules
