+++ bitcoin-core-sv2/src/unix_capnp/mod.rs
@@ -9,6 +9,8 @@
 
 pub mod v30x;
 pub mod v31x;
+pub mod v32x;
 
 // Shared cross-version implementation modules are internal plumbing, not stable public API.
-pub(crate) mod v31x_v30x;
+pub(crate) mod v32x_v31x;
+pub(crate) mod v32x_v31x_v30x;
