+++ bitcoin-core-sv2/src/unix_capnp/v31x/mod.rs
@@ -6,5 +6,7 @@
 //! It is wired against `bitcoin_capnp_types_v31`, which re-exports the matching `capnp`
 //! and `capnp-rpc` APIs.
 
+pub(crate) use bitcoin_capnp_types_v31 as bitcoin_capnp_types;
+
 pub mod job_declaration_protocol;
 pub mod template_distribution_protocol;
