--- bitcoin-core-sv2/src/unix_capnp/v31x_v30x/job_declaration_protocol/mod.rs
@@ -1,3 +1,0 @@
-//! Shared Job Declaration Protocol modules for Bitcoin Core v30.x and v31.x.
-
-pub mod mempool;
