+++ bitcoin-core-sv2/src/unix_capnp/v32x_v31x/job_declaration_protocol/handlers.rs
@@ -1,9 +1,7 @@
-//! Handlers for Bitcoin Core v30.x Sv2 Job Declaration Protocol via capnp over UNIX socket.
+//! Shared JDP handlers for Bitcoin Core v31.x and v32.x runtimes.
 
-use crate::{
-    runtime_api::job_declaration_protocol::io::{JdResponse, ValidationContext},
-    unix_capnp::v30x::job_declaration_protocol::BitcoinCoreSv2JDP,
-};
+use super::BitcoinCoreSv2JDP;
+use crate::runtime_api::job_declaration_protocol::io::{JdResponse, ValidationContext};
 use stratum_core::{
     bitcoin::{
         Block, Transaction, TxMerkleNode, Wtxid,
@@ -23,9 +21,9 @@ impl BitcoinCoreSv2JDP {
     /// Validates a declared mining job by checking transaction availability and block structure.
     ///
     /// Adds missing transactions to the mempool mirror, verifies all transactions are available,
-    /// assembles a test block, and uses Bitcoin Core's `checkBlock` to validate the block
-    /// structure. Returns success with current template parameters or an error if validation
-    /// fails.
+    /// assembles a test block, sets IPC thread context, and uses Bitcoin Core's `checkBlock` to
+    /// validate the block structure. Returns success with current template parameters or an error
+    /// if validation fails.
     pub(crate) async fn handle_declare_mining_job(
         &self,
         version: Version,
@@ -124,11 +122,26 @@ impl BitcoinCoreSv2JDP {
             );
 
             let mut check_block_request = self.mining_ipc_client.check_block_request();
-            let mut check_block_params = check_block_request.get();
 
-            check_block_params.set_block(&block_bytes);
+            match check_block_request.get().get_context() {
+                Ok(mut context) => context.set_thread(self.thread_ipc_client.clone()),
+                Err(e) => {
+                    error!("Failed to set check block request thread context: {e}");
+                    // send error response to the client
+                    // deliberately ignore potential send errors
+                    let _ = response_tx.send(JdResponse::Error {
+                        error_code: ERROR_CODE_DECLARE_MINING_JOB_INTERNAL_ERROR,
+                        validation_context: initial_validation_context,
+                    });
+                    warn!("Terminating Sv2 Bitcoin Core IPC Connection");
+                    self.cancellation_token.cancel();
+                    return;
+                }
+            }
 
-            let mut options = match check_block_params.get_options() {
+            check_block_request.get().set_block(&block_bytes);
+
+            let mut options = match check_block_request.get().get_options() {
                 Ok(options) => options,
                 Err(e) => {
                     error!("Failed to get check block options: {e}");
@@ -275,12 +288,5 @@ impl BitcoinCoreSv2JDP {
         // deliberately ignore potential send errors
         // we don't care if the receiver dropped the channel
         let _ = response_tx.send(response);
-    }
-
-    /// Submits a solved block to Bitcoin Core.
-    ///
-    /// Not yet implemented for v30.x IPC, which does not expose `submitBlock`.
-    pub(crate) async fn handle_push_solution(&self, _block: Block) {
-        // todo
     }
 }
