+++ bitcoin-core-sv2/src/unix_capnp/v32x_v31x_v30x/job_declaration_protocol/mod.rs
@@ -1,0 +1,3 @@
+//! Shared Job Declaration Protocol modules for Bitcoin Core v30.x, v31.x, and v32.x.
+
+pub mod mempool;
